requirements.txt
yfinance
pandas
numpy
PyYAML
pytest
pytest-mock
